Remove assertion on irrelevant TX
Due to how bloom filters work (they give false
positives) it's meaningless to assert on _not_
receiving a transaction from our peer. In most
cases we're not going to receive an irrelevant
TX, but if it happens we can't be sure if it's
because we did something wrong or because we
just hit a false positive.